make primary entity optional for observation table creation for now
